add stealing from santa

+++ refresh-submodules.sh
@@ -2,11 +2,22 @@
 
 git submodule update --remote --merge
 
-(cd submodules/metamorph && ./build.sh)
-rm -rf sites/metamorph
-mkdir -p sites/metamorph
-mv submodules/metamorph/build sites/metamorph/dist
+if [ "$1" == "" ] || [ "$1" == "metamorph" ] ; then 
+  (cd submodules/metamorph && ./build.sh)
+  rm -rf sites/metamorph
+  mkdir -p sites/metamorph
+  cp -r submodules/metamorph/build sites/metamorph/dist
+fi
 
-rm -rf sites/resume
-mkdir -p sites/resume
-mv submodules/resume/dist sites/resume
+if [ "$1" == "" ] || [ "$1" == "resume" ] ; then
+  rm -rf sites/resume
+  mkdir -p sites/resume
+  cp -r submodules/resume/dist sites/resume
+fi
+
+if [ "$1" == "" ] || [ "$1" == "stealing-from-santa" ] ; then
+  (cd submodules/stealing-from-santa && ./build.sh)
+  rm -rf sites/stealing-from-santa
+  mkdir -p sites/stealing-from-santa
+  cp -r submodules/stealing-from-santa/web sites/stealing-from-santa/dist
+fi
